requirements.txt
Appium-Python-Client==1.0.2
selenium==3.141.0
urllib3==1.25.10
